pyproject.toml
[project]
name = "personal-race-engineer"
version = "0.1.0"
description = "Personal racing engineer for iRacing — telemetry analysis and AI coaching"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.84.0",
    "httpx>=0.28.1",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "scipy>=1.17.1",
    "streamlit>=1.54.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
